requirements.txt
# Core
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
python-dotenv==1.0.0

# Database
aiosqlite>=0.20.0

# Redis
redis>=5.2.0

# MinIO
minio>=7.2.0

# ChromaDB
chromadb>=0.5.0

# Kafka
kafka-python>=2.0.2

# Celery
celery>=5.4.0

# OpenAI
openai>=1.54.0

# Google Gemini
google-generativeai>=0.8.0

# PDF Processing
pdfminer.six>=20221105
pytesseract>=0.3.10

# Stock Analysis
pandas>=2.2.0
numpy>=1.26.0

# Workflow
langchain>=0.3.0
langgraph>=0.2.0

# GitHub
GitPython>=3.1.40

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.21.0

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
hypothesis>=6.100.0
